+++ infrastructor-cli/src/main/java/io/infrastructor/cli/handlers/EncryptHandler.groovy
@@ -6,7 +6,7 @@ import groovy.time.TimeCategory
 import io.infrastructor.cli.validation.FileValidator
 import io.infrastructor.cli.validation.ModeValidator
 
-import static io.infrastructor.core.utils.CryptoUtils2.*
+import static io.infrastructor.core.utils.CryptoUtils.*
 import static io.infrastructor.cli.validation.ModeValidator.*
 import static io.infrastructor.core.logging.ConsoleLogger.*
 import static io.infrastructor.core.logging.status.TextStatusLogger.withTextStatus
