+++ infrastructor-core/src/test/java/io/infrastructor/core/utils/CryptoUtilsTest.groovy
@@ -2,21 +2,21 @@ package io.infrastructor.core.utils
 
 import org.junit.Test
 
-import static io.infrastructor.core.utils.CryptoUtils2.ENCODING
+import static CryptoUtils.ENCODING
 
-class CryptoUtils2Test {
+class CryptoUtilsTest {
 
     final String DATA = "secret message"
     final String KEY  = "secret"
 
     @Test
     void fullEncryptionAndDecryption() {
-        def encrypted = CryptoUtils2.encryptFull(KEY, DATA.getBytes(ENCODING))
+        def encrypted = CryptoUtils.encryptFull(KEY, DATA.getBytes(ENCODING))
         println encrypted
 
         assert encrypted.startsWith('INFRASTRUCTOR:AES/GCM/PKCS5Padding:BASE64:OIEhnQh92cY0Nz/TPfozostr/GxSC2S4u2DvLOtTSuc=:')
 
-        byte [] decrypted = CryptoUtils2.decryptFull(KEY, encrypted)
+        byte [] decrypted = CryptoUtils.decryptFull(KEY, encrypted)
         assert DATA == new String(decrypted, ENCODING)
     }
 
@@ -25,14 +25,14 @@ class CryptoUtils2Test {
         def template = '''
             message: "${decrypt('Xi4Hwkd4apQSlRnAX/iJpNMRofIxay7rmhGWiQmc', '8WJJgQSAyHaFD3Ot')}"
         '''
-        assert CryptoUtils2.decryptPart(KEY, template).contains(DATA)
+        assert CryptoUtils.decryptPart(KEY, template).contains(DATA)
     }
 
     @Test
     void encryptPart() {
-        def template = CryptoUtils2.encryptPart(KEY, """
+        def template = CryptoUtils.encryptPart(KEY, """
             message: "\${encrypt('$DATA')}"
         """)
-        assert CryptoUtils2.decryptPart(KEY, template).contains(DATA)
+        assert CryptoUtils.decryptPart(KEY, template).contains(DATA)
     }
 }
